requirements.txt
requests==2.32.3
bs4
ollama==0.2.1
regex==2023.10.3
joblib==1.3.2
streamlit==1.33.0
pandas==2.1.2
urllib3==2.2.2
python-dotenv==1.0.0
scikit-learn==1.4.1.post1
ollama==0.2.1
Flask==3.0.3
stripe==10.5.0
uvicorn==0.30.1
streamlit-authenticator==0.3.2